+++ Changelog.md
@@ -39,3 +39,4 @@
 - [TechReborn Quantum Suit now requires the Tech Enhanced Nano Suit](https://github.com/TheStaticVoid/StaTech-Industry/issues/407)
 - [Updated Doped silicon plate textures](https://github.com/TheStaticVoid/StaTech-Industry/issues/409)
 - [Buffed the Enderman mob crusher model](https://github.com/TheStaticVoid/StaTech-Industry/issues/414)
+- [Add mixer recipe for Cultural Delights' pickle](https://github.com/TheStaticVoid/StaTech-Industry/issues/410)